databricks-ecommerce-dlt-pipeline/pyproject.toml
[project]
name = "ecommerce-dlt"
version = "0.1.0"
description = "E-Commerce DLT Pipeline: Declarative Delta Live Tables on Olist Brazilian E-Commerce data"
requires-python = ">=3.10"

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "ruff>=0.1",
    "databricks-sdk>=0.20",
    "databricks-connect>=15.4,<15.5",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]
